+++ packages/react-composites/src/composites/CallComposite/CallControls.tsx
@@ -9,7 +9,7 @@ import {
   groupCallLeaveButtonStyle
 } from './styles/CallControls.styles';
 import { usePropsFor } from './hooks/usePropsFor';
-import { devicePermissionSelector } from '@azure/calling-component-bindings';
+import { devicePermissionSelector } from 'calling-component-bindings';
 import { useAdaptedSelector } from './hooks/useAdaptedSelector';
 
 export type GroupCallControlsProps = {