+++ packages/react-composites/src/composites/ChatComposite/hooks/useHandlers.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import { CommonProperties } from 'acs-ui-common';
-import { DefaultChatHandlers } from '@azure/chat-component-bindings';
+import { DefaultChatHandlers } from 'chat-component-bindings';
 
 import { ReactElement } from 'react';
 import memoizeOne from 'memoize-one';