+++ samples/Calling/src/app/CallControls.tsx
@@ -3,13 +3,13 @@
 
 import React, { useCallback } from 'react';
 import { ControlBar, MicrophoneButton, CameraButton, ScreenShareButton, EndCallButton } from 'react-components';
-import { devicePermissionSelector, useCallingPropsFor as usePropsFor } from '@azure/calling-component-bindings';
+import { devicePermissionSelector, useCallingPropsFor as usePropsFor } from 'calling-component-bindings';
 import {
   controlBarStyle,
   groupCallLeaveButtonCompressedStyle,
   groupCallLeaveButtonStyle
 } from './styles/CallControls.styles';
-import { useCallingSelector as useSelector } from '@azure/calling-component-bindings';
+import { useCallingSelector as useSelector } from 'calling-component-bindings';
 
 export type CallControlsProps = {
   onEndCallClick(): void;