+++ samples/Calling/src/app/MediaGallery.tsx
@@ -3,10 +3,7 @@
 
 import React, { useEffect, useMemo, useState } from 'react';
 import { VideoGallery, VideoStreamOptions } from 'react-components';
-import {
-  useCallingPropsFor as usePropsFor,
-  useCallingSelector as useSelector
-} from '@azure/calling-component-bindings';
+import { useCallingPropsFor as usePropsFor, useCallingSelector as useSelector } from 'calling-component-bindings';
 import { ScreenShare } from './ScreenShare';
 import { getIsPreviewCameraOn } from './selectors/baseSelectors';