+++ samples/Calling/src/app/hooks/useAzureCommunicationHandlers.ts
@@ -7,7 +7,7 @@ import {
   useCallAgent,
   useCallClient,
   useDeviceManager
-} from '@azure/calling-component-bindings';
+} from 'calling-component-bindings';
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types, @typescript-eslint/explicit-function-return-type
 export const useAzureCommunicationHandlers = () => {