+++ samples/Chat/src/app/ChatArea.tsx
@@ -5,7 +5,7 @@ import { SendBox, TypingIndicator, MessageThread } from 'react-components';
 import { Stack } from '@fluentui/react';
 import React, { useEffect } from 'react';
 import { chatAreaContainerStyle, sendBoxParentStyle } from './styles/ChatArea.styles';
-import { useChatPropsFor as usePropsFor } from '@azure/chat-component-bindings';
+import { useChatPropsFor as usePropsFor } from 'chat-component-bindings';
 
 export interface ChatAreaProps {
   onRenderAvatar?: (userId: string) => JSX.Element;